Gird.java -Used Arrays.Streams - its a bit better for performance
GridWorldInfo.java
changed runtime to double so we could do this in seconds

Test.java
Updated to print s
added color constants

Robot.java
-Changed runtime calc to divide by nanoseconds

NaiveAgent.java
-Made can see sideways non static

InfrenceAgent
-Added Default methods to make the propogate a bit easier

+++ com.rutgers.asearch/src/Entity/GridWorldInfo.java
@@ -11,7 +11,7 @@ public class GridWorldInfo {
     public int numBumps;
     public int numPlans;
     public int numCellsDetermined;
-    public long runtime;
+    public double runtime;
 
     public List<Point> path; // path does not include start cell
